Criado novo evento e consertado conversão
Evento OnClienteIndisponivel indica quando um cliente, que estava na
rede, saiu.

Consertada conversão de Direcao

+++ BatalhaNaval/Tabuleiro.cs
@@ -259,7 +259,7 @@ namespace BatalhaNaval
             foreach (Celula celula in celulas)
                 this[celula.Linha, celula.Coluna] = celula;
             
-            Navios.Add(new int[] { x, y, d }, tipo);
+            Navios.Add(new int[] { x, y, (int) d }, tipo);
         }
 
         /// <summary>
